formal/promela/src/requirements.txt
astroid==2.4.2
attrs==20.3.0
coconut==1.4.3
coverage==5.3
cPyparsing==2.4.5.0.1.1
decorator==4.4.2
flake8==3.8.4
importlib-metadata==3.1.1
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
networkx==2.5
packaging==20.7
parsec==3.5
pluggy==0.13.1
ply==3.10
prompt-toolkit==3.0.8
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.6.0
pytest==6.1.2
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0
